requirements.txt
hydra-core==1.3.2
numpy==1.24.4
joblib==1.4.2
einops==0.8.0
matplotlib==3.7.5
shapely==2.0.6
scipy==1.10.1
opencv-python==4.10.0.84
imageio==2.34.2